pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"

[tool.poetry]
name = "gans_zoo"
version = "0.1.0"
description = ""
authors = ["Ilya Siganov <ilya.blan4@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/senior-sigan/gans_zoo"

[tool.poetry.dependencies]
python = "3.8.5"
torch = "^1.7.0"
torchvision = "^0.8.1"
pytorch-lightning = "^1.0.7"
pytorch-lightning-bolts = "^0.2.5"
python-telegram-bot = "^13.0"

[tool.poetry.dev-dependencies]
mypy = "^0.790"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.3"
nitpick = "^0.23"

pytest = "^5.4"
pytest-cov = "^2.10"
pytest-randomly = "^3.5"

jupyter = "^1.0.0"
ipython = "^7.19.0"
autopep8 = "^1.5.4"
pytest-dotenv = "^0.5.2"
